certificate-importer/requirements.txt
dnspython==2.8.0
numpy==2.3.3
pandas==2.3.2
pymongo==4.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
tzdata==2025.2
